refactor: auth dtos
