--- server/src/domain/auth/dto/change-password.dto.ts
@@ -1,15 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsNotEmpty, IsString, MinLength } from 'class-validator';
-
-export class ChangePasswordDto {
-  @IsString()
-  @IsNotEmpty()
-  @ApiProperty({ example: 'password' })
-  password!: string;
-
-  @IsString()
-  @IsNotEmpty()
-  @MinLength(8)
-  @ApiProperty({ example: 'password' })
-  newPassword!: string;
-}
