--- server/src/domain/auth/dto/index.ts
@@ -1,6 +1,0 @@
-export * from './auth-user.dto';
-export * from './change-password.dto';
-export * from './login-credential.dto';
-export * from './oauth-auth-code.dto';
-export * from './oauth-config.dto';
-export * from './sign-up.dto';
