--- server/src/domain/auth/dto/oauth-auth-code.dto.ts
@@ -1,9 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class OAuthCallbackDto {
-  @IsNotEmpty()
-  @IsString()
-  @ApiProperty()
-  url!: string;
-}
