--- server/src/domain/auth/dto/oauth-config.dto.ts
@@ -1,7 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class OAuthConfigDto {
-  @IsNotEmpty()
-  @IsString()
-  redirectUri!: string;
-}
