+++ server/src/domain/auth/index.ts
@@ -1,4 +1,3 @@
 export * from './auth.constant';
+export * from './auth.dto';
 export * from './auth.service';
-export * from './dto';
-export * from './response-dto';
