--- server/src/domain/auth/response-dto/index.ts
@@ -1,6 +1,0 @@
-export * from './admin-signup-response.dto';
-export * from './auth-device-response.dto';
-export * from './login-response.dto';
-export * from './logout-response.dto';
-export * from './oauth-config-response.dto';
-export * from './validate-asset-token-response.dto';
