--- server/src/domain/auth/response-dto/logout-response.dto.ts
@@ -1,4 +1,0 @@
-export class LogoutResponseDto {
-  successful!: boolean;
-  redirectUri!: string;
-}
