--- server/src/domain/auth/response-dto/oauth-config-response.dto.ts
@@ -1,11 +1,0 @@
-export class OAuthConfigResponseDto {
-  enabled!: boolean;
-  passwordLoginEnabled!: boolean;
-  url?: string;
-  buttonText?: string;
-  autoLaunch?: boolean;
-}
-
-export class OAuthAuthorizeResponseDto {
-  url!: string;
-}
