--- server/src/domain/auth/response-dto/validate-asset-token-response.dto.ts
@@ -1,3 +1,0 @@
-export class ValidateAccessTokenResponseDto {
-  authStatus!: boolean;
-}
