+++ server/test/e2e/auth.e2e-spec.ts
@@ -8,7 +8,6 @@ import {
   errorStub,
   loginResponseStub,
   loginStub,
-  signupResponseStub,
   uuidStub,
 } from '@test/fixtures';
 import { testApp } from '@test/test-utils';
@@ -18,6 +17,24 @@ const firstName = 'Immich';
 const lastName = 'Admin';
 const password = 'Password123';
 const email = 'admin@immich.app';
+
+const adminSignupResponse = {
+  id: expect.any(String),
+  firstName: 'Immich',
+  lastName: 'Admin',
+  email: 'admin@immich.app',
+  storageLabel: 'admin',
+  externalPath: null,
+  profileImagePath: '',
+  // why? lol
+  shouldChangePassword: true,
+  isAdmin: true,
+  createdAt: expect.any(String),
+  updatedAt: expect.any(String),
+  deletedAt: null,
+  oauthId: '',
+  memoriesEnabled: true,
+};
 
 describe(`${AuthController.name} (e2e)`, () => {
   let server: any;
@@ -84,7 +101,7 @@ describe(`${AuthController.name} (e2e)`, () => {
         .post('/auth/admin-sign-up')
         .send({ ...adminSignupStub, email: 'admin@local' });
       expect(status).toEqual(201);
-      expect(body).toEqual({ ...signupResponseStub, email: 'admin@local' });
+      expect(body).toEqual({ ...adminSignupResponse, email: 'admin@local' });
     });
 
     it('should transform email to lower case', async () => {
@@ -92,7 +109,7 @@ describe(`${AuthController.name} (e2e)`, () => {
         .post('/auth/admin-sign-up')
         .send({ ...adminSignupStub, email: 'aDmIn@IMMICH.app' });
       expect(status).toEqual(201);
-      expect(body).toEqual(signupResponseStub);
+      expect(body).toEqual(adminSignupResponse);
     });
 
     it('should not allow a second admin to sign up', async () => {
