+++ server/test/fixtures/auth.stub.ts
@@ -12,14 +12,6 @@ export const userSignupStub = {
   memoriesEnabled: true,
 };
 
-export const signupResponseStub = {
-  id: expect.any(String),
-  email: 'admin@immich.app',
-  firstName: 'Immich',
-  lastName: 'Admin',
-  createdAt: expect.any(String),
-};
-
 export const loginStub = {
   admin: {
     email: 'admin@immich.app',
